Separate RecipeCard to reusable view
